相机测距及参数配置工具/yaml读取/distance_yaml_0.0.2/requirements.txt
numpy~=1.19.3
opencv-python~=4.5.3.56
pillow~=9.2.0
setuptools~=61.2.0
cython~=0.29.21
pycocotools~=2.0.4
pyyaml~=5.3.1